requirements.txt
discord.py
pandas
openpyxl
python-dotenv